requirements-pip3.txt
psycopg2==2.9.7
pandas==1.5.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
streamlit==1.25.0
streamlit-lottie
python-dotenv
